Blaze Defconfig: Enable CONFIG_MODVERSIONS
CONFIG_MODVERSIONS allows modules to be used
with a different kernel than what it was built
against. When CONFIG_MODVERSIONS is enabled,
the kernel will add a special piece of versioning
info to every exported symbol. This versioning info
is calculated by a CRC algorithm which recursively
looks at the symbols associated structs, unions,
enums, and typedefs. When an external module is built,
it will store the corresponding symbol and version info.
When the module is inserted a check is performed
to ensure that the version for each symbol matches.

This ensures that the interface between kernel
and modules is not broken, and allows more flexibility
than an explicit kernel version check.

Change-Id: Idd358dcaa86f1c77544944899d1c3d355d6c15e6
Signed-off-by: Tate Hornbeck <tate.hornbeck@ti.com>